requirements.txt
flask
azure.cosmos
pandas
numpy
dataclasses